Separated logic.identity
