+++ src/main/java/gui/controller/EmployeeController.java
@@ -16,8 +16,8 @@ import gui.helper.EmployeeManagerWrapper;
 import gui.helper.EmployeeRow;
 import javafx.fxml.FXML;
 import javafx.scene.text.Text;
-import logic.human.ID;
-import logic.human.IDType;
+import logic.identity.ID;
+import logic.identity.IDType;
 import logic.human.NameBuilder;
 import logic.human.Sex;
 import logic.money.Money;
